het loi ne

+++ server/src/domains/users/users.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   Delete,
   Request,
+  HttpException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -17,7 +18,7 @@ export class UsersController {
 
   @Patch()
   update(@Request() req: any, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(req.user.uid, updateUserDto);
+    return this.usersService.update(req.user.uid || req.user.id, updateUserDto);
   }
 
   @Post()
@@ -38,7 +39,11 @@ export class UsersController {
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string) {
-    return await this.usersService.remove(id);
+  async remove(@Param('id') id: string, @Request() req: any) {
+    if (req.user.role) {
+      return await this.usersService.remove(id);
+    } else {
+      throw new HttpException('Permission denied', 403);
+    }
   }
 }
